+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add into_group() method to GroupExt.
 - Pull fixes from FLTK and cfltk.
 - Update changelog.
+- Allow set_highlight_data to be set even in the absence of a text buffer.
 
 ## [1.0.20] - 2021-06-26
 - Add app::event_original_key().