+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Pull fixes from FLTK and cfltk.
 - Update changelog.
 - Allow set_highlight_data to be set even in the absence of a text buffer.
+- Add MenuButton types enum MenuButtonType.
 
 ## [1.0.20] - 2021-06-26
 - Add app::event_original_key().